requirements.txt
autopep8==1.6.0
imutils==0.5.4
numpy==1.22.3
opencv-python==4.5.5.64
pycodestyle==2.8.0
python-dotenv==0.19.2
scipy==1.8.0
toml==0.10.2
